refactor: move button core and base CSS to styles folder
